requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.10.4
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==2.0.12
click==8.1.7
colorama==0.4.6
deprecation==2.1.0
exceptiongroup==1.2.2
Flask==2.2.5
frozenlist==1.4.1
gotrue==2.9.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
multidict==6.1.0
packaging==24.1
postgrest==0.17.1
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==0.19.2
python-telegram-bot==21.6
pytz==2024.2
realtime==2.0.5
requests==2.26.0
six==1.16.0
sniffio==1.3.1
storage3==0.8.1
StrEnum==0.4.15
supabase==2.9.0
supafunc==0.6.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==1.26.20
websockets==13.1
Werkzeug==2.2.3
yarl==1.15.4
zipp==3.20.2
